smart header wip
