+++ src/cmps/layout/HeaderStayDetails.jsx
@@ -1,0 +1,289 @@
+import { Link, NavLink, useSearchParams } from "react-router-dom";
+import { useNavigate } from "react-router";
+import { useSelector } from "react-redux";
+import { showErrorMsg, showSuccessMsg } from "../../services/event-bus.service";
+import { userActions } from "../../store/actions/user.actions";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faBars } from "@fortawesome/free-solid-svg-icons";
+import { SearchBar } from "../stay/SearchBar";
+import { useEffect, useRef, useState } from "react";
+import { stayActions } from "../../store/actions/stay.actions";
+import { animateCSS, getExistingProperties } from "../../services/util.service";
+import { SearchBarMini } from "../stay/SearchBarMini";
+import logo from "../../assets/img/logo.png";
+import guestUnknown from "../../assets/img/guest-unknown.svg";
+import { stayService } from "../../services/stay";
+import { addDays } from "date-fns";
+import { NavMenu } from "../stay/NavMenu";
+
+export function HeaderStayDetails({ showAnchor }) {
+  const user = useSelector((storeState) => storeState.userModule.user);
+  const [searchParams, setSearchParams] = useSearchParams();
+  const [activeSearchControl, setActiveSearchControl] = useState("");
+  const city = useSelector((storeState) => storeState.stayModule.filterBy.city);
+  const [destination, setDestination] = useState(city || "");
+  const datesRange = useSelector(
+    (storeState) => storeState.stayModule.datesRange
+  );
+  const guests = useSelector((storeState) => storeState.stayModule.guests);
+  const [guestsDisplay, setGuestsDisplay] = useState("");
+  const [isSearchBarVisible, setIsSearchBarVisible] = useState(false);
+  // const [isAtTop, setIsAtTop] = useState(() => window.scrollY === 0);
+  // const [isManuallyTriggered, setIsManuallyTriggered] = useState(false);
+  const [isNavMenuVisible, setIsNavMenuVisible] = useState(false);
+  // const isManuallyTriggeredRef = useRef(false);
+  // const justTriggeredManually = useRef(false);
+  // const shouldShowSearchBar = isAtTop || isManuallyTriggered;
+  const searchBarRef = useRef();
+  const miniSearchBarRef = useRef();
+  const homesTitleRef = useRef();
+  // const navigate = useNavigate();
+
+  useEffect(() => {
+    stayActions.setFilterBy(
+      stayService.getFilterByFromSearchParams(searchParams)
+    );
+    stayActions.setGuests(stayService.getGuestsFromSearchParams(searchParams));
+    stayActions.setDatesRange(
+      stayService.getDatesRangeFromSearchParams(searchParams)
+    );
+  }, [searchParams]);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setIsSearchBarVisible(false);
+    };
+
+    window.addEventListener("scroll", handleScroll);
+
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, []);
+
+  // useEffect(() => {
+  //   const handleScroll = () => {
+  //     const atTop = window.scrollY === 0;
+  //     setIsAtTop(atTop);
+
+  //     if (
+  //       !atTop &&
+  //       isManuallyTriggeredRef.current &&
+  //       !justTriggeredManually.current
+  //     ) {
+  //       setIsManuallyTriggered(false);
+  //       isManuallyTriggeredRef.current = false;
+  //     }
+  //   };
+
+  //   window.addEventListener("scroll", handleScroll);
+  //   handleScroll(); // initialize
+  //   return () => window.removeEventListener("scroll", handleScroll);
+  // }, []);
+
+  // useEffect(() => {
+  //   if (isAtTop) {
+  //     setIsManuallyTriggered(false); // clear manual override when back at top
+  //   }
+  // }, [isAtTop]);
+
+  // useEffect(() => {
+  //   if (shouldShowSearchBar) {
+  //     animateSearchBarEntrance();
+  //   } else {
+  //     animateSearchBarLeave();
+  //   }
+  //   async function animateSearchBarEntrance() {
+  //     await setIsSearchBarVisible(true);
+  //     animateCSS(searchBarRef.current, "fadeInDown");
+  //     animateCSS(homesTitleRef.current, "fadeInDown");
+  //   }
+
+  //   async function animateSearchBarLeave() {
+  //     setIsSearchBarVisible(false);
+  //   }
+  // }, [shouldShowSearchBar]);
+
+  useEffect(() => {
+    if (!isSearchBarVisible) {
+      animateCSS(miniSearchBarRef.current, "fadeInUp");
+    } else {
+      animateCSS(searchBarRef.current, "fadeInDown");
+      animateCSS(homesTitleRef.current, "fadeInDown");
+    }
+  }, [isSearchBarVisible]);
+
+  useEffect(() => {
+    setDestination(city);
+  }, [city]);
+
+  useEffect(() => {
+    setGuestsDisplay(renderGuestsDisplay());
+  }, [guests]);
+
+  function onSetDatesRange(datesRange) {
+    stayActions.setDatesRange(datesRange);
+  }
+
+  function renderGuestsDisplay() {
+    const guestsDisplayArray = [];
+    let guestCount = 0;
+    guests.forEach(({ type, count }) => {
+      switch (type) {
+        case "adults":
+          guestCount += count;
+          break;
+        case "children":
+          guestCount += count;
+          break;
+        case "infants":
+          {
+            if (count > 0) {
+              guestsDisplayArray.push(`${count} infant${count > 1 ? "s" : ""}`);
+            }
+          }
+          break;
+        case "pets":
+          {
+            if (count > 0) {
+              guestsDisplayArray.push(`${count} pet${count > 1 ? "s" : ""}`);
+            }
+          }
+          break;
+        default:
+          break;
+      }
+    });
+
+    if (guestCount > 0)
+      guestsDisplayArray.unshift(
+        `${guestCount} guest${guestCount > 1 ? "s" : ""}`
+      );
+
+    return !guestsDisplayArray.length
+      ? "Add guests"
+      : guestsDisplayArray.join(", ");
+  }
+
+  function onSetGuests(newGuests) {
+    stayActions.setGuests([...newGuests]);
+  }
+
+  async function updateFilterBy() {
+    let endDate = datesRange[1];
+
+    if (datesRange[0] && !datesRange[1]) {
+      endDate = addDays(datesRange[0], 1);
+      await stayActions.setDatesRange([datesRange[0], endDate]);
+    }
+
+    const capacity = guests.reduce((acc, guest) => acc + guest.count, 0);
+
+    const newFilterBy = {
+      city: destination,
+      startDate: datesRange[0],
+      endDate,
+      capacity,
+    };
+
+    const isPetsAllowed = guests.some(
+      (guest) => guest.type === "pets" && guest.count > 0
+    );
+
+    if (isPetsAllowed) {
+      newFilterBy.isPetsAllowed = isPetsAllowed;
+    }
+
+    guests.forEach((guestObj) => {
+      newFilterBy[guestObj.type] = guestObj.count;
+    });
+
+    const type = searchParams.get("type");
+    if (type) {
+      newFilterBy.type = type;
+    }
+
+    setSearchParams({ ...getExistingProperties(newFilterBy) });
+  }
+
+  function handleMiniSearchBarClick(controlType) {
+    setIsSearchBarVisible(true);
+    setActiveSearchControl(controlType);
+    // setIsManuallyTriggered(true);
+    // isManuallyTriggeredRef.current = true;
+    // justTriggeredManually.current = true;
+    // setTimeout(() => {
+    //   justTriggeredManually.current = false;
+    // }, 100);
+    // setActiveSearchControl(controlType);
+  }
+  // async function onLogout() {
+  //   try {
+  //     await logout();
+  //     navigate("/");
+  //     showSuccessMsg(`Bye now`);
+  //   } catch (err) {
+  //     showErrorMsg("Cannot logout");
+  //   }
+  // }
+
+  function handleUserIconClick() {
+    if (isNavMenuVisible) {
+      setIsNavMenuVisible(false);
+    } else {
+      setIsNavMenuVisible(true);
+    }
+  }
+
+  const mainHeader = (
+    <>
+      <NavLink className={"logo-link"} to={""}>
+        <div className="logo-container">
+          <img className="logo" src={logo} alt="logo" />
+          <h3>kmobnb</h3>
+        </div>
+      </NavLink>
+      {isSearchBarVisible ? (
+        <div ref={homesTitleRef} className="homes-title-container">
+          <h2>Homes</h2>
+        </div>
+      ) : (
+        <div ref={miniSearchBarRef} className="mini-search-bar-container">
+          <SearchBarMini onSelect={handleMiniSearchBarClick} />
+        </div>
+      )}
+      <nav>
+        <NavLink>Bnb your home</NavLink>
+        <button className="user-info" onClick={handleUserIconClick}>
+          <FontAwesomeIcon icon={faBars} />
+          <img src={user?.imgUrl ?? guestUnknown} alt="user-icon" />
+        </button>
+        {isNavMenuVisible ? <NavMenu /> : ""}
+      </nav>
+    </>
+  );
+
+  const secondaryHeader = <></>;
+
+  return (
+    <>
+      <section className="header-top">
+        {showAnchor ? secondaryHeader : mainHeader}
+      </section>
+      {isSearchBarVisible && (
+        <div ref={searchBarRef}>
+          <SearchBar
+            activeSearchControl={activeSearchControl}
+            setActiveSearchControl={setActiveSearchControl}
+            destination={destination}
+            setDestination={setDestination}
+            datesRange={datesRange}
+            onSetDatesRange={onSetDatesRange}
+            guests={guests}
+            guestsDisplay={guestsDisplay}
+            onSetGuests={onSetGuests}
+            updateFilterBy={updateFilterBy}
+          />
+        </div>
+      )}
+    </>
+  );
+}
